umbrella_core: prevent VR lag by decreasing load limit
The original version of the governor focused on attempting to provide
the best performance at the least drain, but this has caused some
slight jumps and lags when using the Gear VR.

Since this kernel targets users with a VR headset and an extended
battery, it is not longer a priority to focus on battery at the cost of
VR performance. Both are equal.